requirements.txt
absl-py==0.13.0
alembic==1.7.3
anyio==3.3.1
apache-airflow==2.1.4
apache-airflow-providers-amazon==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.2.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
astunparse==1.6.3
atomicwrites==1.4.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
boto3==1.17.112
botocore==1.20.112
cachetools==4.2.2
category-encoders==2.2.2
cattrs==1.5.0
certifi==2020.12.5
charset-normalizer==2.0.6
clang==5.0
click==7.1.2
clickclick==20.10.2
colorlog==5.0.1
commonmark==0.9.1
coverage==5.5
croniter==1.0.15
cryptography==3.4.8
cycler==0.10.0
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.17
grpcio==1.40.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
importlib-resources==1.5.0
inflection==0.5.1
iniconfig==1.1.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
jmespath==0.10.0
joblib==1.0.1
jsonpath-ng==1.5.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.1
mistune==0.8.4
numpy==1.19.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.2.4
paramiko==2.7.2
patsy==0.5.1
pendulum==2.1.2
Pillow==8.2.0
pluggy==0.13.1
ply==3.11
pprintpp==0.4.0
prison==0.2.1
protobuf==3.18.0
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycountry-convert==0.7.2
PyJWT==1.7.1
PyNaCl==1.4.0
pysftp==0.2.9
pytest==6.2.3
pytest-cov==2.11.1
pytest-mock==3.5.1
python-daemon==2.3.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
repoze.lru==0.7
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.10.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
setproctitle==1.2.2
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sshtunnel==0.4.0
statsmodels==0.12.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==6.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.6
watchtower==1.0.6
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
